// packages/@oxen-office/pptx/src/domain/diagram/types.ts
export type DiagramBoolOperator = "equ" | "gte" | "lte" | "none";
// ...
export type DiagramConstraintRelationship = "ch" | "des" | "self";

export type DiagramConstraintType =
  | "alignOff"
  | "b"
  | "begMarg"
  | "begPad"
  | "bendDist"
  | "bMarg"
  | "bOff"
  | "connDist"
  | "ctrX"
  | "ctrXOff"
  | "ctrY"
  | "ctrYOff"
  | "diam"
  | "endMarg"
  | "endPad"
  | "h"
  | "hArH"
  | "hOff"
  | "l"
  | "lMarg"
  | "lOff"
  | "none"
  | "primFontSz"
  | "pyraAcctRatio"
  | "r"
  | "rMarg"
  | "rOff"
  | "secFontSz"
  | "secSibSp"
  | "sibSp"
  | "sp"
  | "stemThick"
  | "t"
  | "tMarg"
  | "tOff"
  | "userA"
  | "userB"
  | "userC"
  | "userD"
  | "userE"
  | "userF"
  | "userG"
  | "userH"
  | "userI"
  | "userJ"
  | "userK"
  | "userL"
  | "userM"
  | "userN"
  | "userO"
  | "userP"
  | "userQ"
  | "userR"
  | "userS"
  | "userT"
  | "userU"
  | "userV"
  | "userW"
  | "userX"
  | "userY"
  | "userZ"
  | "w"
  | "wArH"
  | "wOff";
// ...
export type DiagramConstraint = {
  readonly type?: DiagramConstraintType;
  readonly forRelationship?: DiagramConstraintRelationship;
  readonly forName?: string;
  readonly referenceType?: DiagramConstraintType;
  readonly referenceForRelationship?: DiagramConstraintRelationship;
  readonly referenceForName?: string;
  readonly operator?: DiagramBoolOperator;
  readonly value?: string;
  readonly factor?: string;
  readonly max?: string;
  readonly min?: string;
};